Ajout d'une grille d'occupation
Ajout de Grid.cpp.h
Modification de PathFinding.cpp.h pour adaptation a la grille
d'occupation

+++ PathFinder/Vector2.cpp
@@ -3,7 +3,7 @@
 #include "Vector2.h"
 
 Vector2::Vector2() : m_x(0), m_y(0) {}
-Vector2::Vector2(int x, int y) : m_x(x), m_y(y) {}
+Vector2::Vector2(float x, float y) : m_x(x), m_y(y) {}
 
 Vector2 Vector2::GetVector()
 {
@@ -11,27 +11,27 @@ Vector2 Vector2::GetVector()
 	return temp;
 }
 
-int Vector2::GetX() const
+float Vector2::GetX() const
 { 
 	return m_x; 
 }
 
-int Vector2::GetY() const
+float Vector2::GetY() const
 { 
 	return m_y; 
 }
 
-void Vector2::SetX(int x)
+void Vector2::SetX(float x)
 { 
 	m_x = x; 
 }
-void Vector2::SetY(int y)
+void Vector2::SetY(float y)
 { 
 	m_y = y; 
 }
 
-int Vector2::Length()
+float Vector2::Length()
 { 
-	return (int)(sqrt((float)(m_x*m_x + m_y*m_y))); 
+	return (sqrt(m_x*m_x + m_y*m_y));
 }
